[Tank.500m] Step1,2 제출입니다. (#6)
* docs: 1단계 기능 명세 추가

* feat : 토큰 발급하는 API 생성

* feat : 토큰 발급하는 API 생성

* docs : 2단계 요구사항 추가

* feat : AuthenticationPrincipalArgumentResolver 추가

* feat : accessToken이 있는 유저만 예약 추가/삭제 가능하도록 수정

* feat : accessToken이 있는 유저만 예약 추가/삭제 가능하도록 수정

* feat : 예약 삭제 시 본인의 예약만 삭제 가능하도록 수정

* style: 포맷팅

---------

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -3,11 +3,13 @@ package nextstep.reservation;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
 import nextstep.support.DuplicateEntityException;
+import nextstep.support.ForbiddenAccessException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class ReservationService {
@@ -49,10 +51,14 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, String username) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
+        }
+
+        if(!Objects.equals(reservation.getName(), username)){
+            throw new ForbiddenAccessException();
         }
 
         reservationDao.deleteById(id);
